More refactoring in the hi-res plots
Rather than needing to know ahead of time which files are archived and which
are in the run directory, get_files() will look on its own (and it accepts a
list of casenames or a single string). I also added a single plot containing
all four lines, though I like the panel plot more.